requirements.txt
aioredis==1.3.1
amqp==5.1.1
aniso8601==7.0.0
asgi-redis==1.4.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.6.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configparser==5.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cssutils==2.5.1
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.6
django-admin-index==2.0.2
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-durationwidget==1.0.5
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.4
django-grappelli==3.0.3
django-hosts==5.1
django-inlinecss==0.3.0
django-ordered-model==3.6
django-phone-field==1.8.1
django-session-cleanup==4.0.0
django-suit==0.2.28
django-timezone-field==5.0
djangorestframework==3.13.1
drf-yasg==1.21.3
flake8==5.0.4
flower==1.2.0
future==0.18.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==2.0.0
humanize==4.3.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
mailchimp-marketing==3.0.75
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
mock==4.0.3
msgpack==1.0.4
msgpack-python==0.5.6
mypy-extensions==0.4.3
mysqlclient==2.1.1
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.4.0
pynliner==0.8.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
tornado==6.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0
